requirements.txt
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
databases[postgresql]==0.7.0
sqlalchemy==1.4.46
celery==5.3.5
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic[email]==2.4.2
python-multipart==0.0.6
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.39.0
python-jose
bcrypt